jz4780: Select RAM device and apply settings from structure
Replace calculation of DDR parameters from header files
with the pre-determined register values in the configuration
structure.
This is not as portable because the settings must be determined
and hard-coded, however, it was necessary to allow selection
of different RAM devices at runtime, and free's up some storage
for additional device parameters if they become necessary.

Signed-off-by: Matt Redfearn <matt.redfearn@imgtec.com>